fix: separate app and server for order-updates and inventory-import to prevent port conflicts
